chore: initial structure for v2
