--- jest.config.js
@@ -1,6 +1,0 @@
-const { createJestConfig } = require("tsdx/dist/createJestConfig");
-const { paths } = require("tsdx/dist/constants");
-const dotenv = require("dotenv");
-dotenv.config();
-
-module.exports = createJestConfig(undefined, paths.appRoot);
