+++ src/old/utils/api-base.ts
@@ -1,6 +1,5 @@
 import axios from "axios";
 import config from "../config";
-import { store } from "../store";
 
 export const baseUrl = config.BASE_URL;
 
@@ -12,13 +11,6 @@ const axiosInstance = axios.create({
     "Access-Control-Allow-Origin": "*",
   },
   responseType: "json",
-});
-
-store.subscribe((state) => {
-  axiosInstance.defaults.auth = {
-    username: state.secretKey,
-    password: "",
-  };
 });
 
 export default axiosInstance;
